move table and column to sub package
