+++ src/java/com/rapleaf/jack/queries/Column.java
@@ -1,4 +1,4 @@
-package com.rapleaf.jack;
+package com.rapleaf.jack.queries;
 
 public class Column {
   private static String DEFAULT_ID_FIELD = "id";
