+++ src/java/com/rapleaf/jack/queries/OrderCriterion.java
@@ -1,7 +1,5 @@
 package com.rapleaf.jack.queries;
 
-import com.rapleaf.jack.Column;
-
 public class OrderCriterion implements QueryCondition {
   private Column column;
   private final QueryOrder order;
