+++ src/java/com/rapleaf/jack/queries/Table.java
@@ -1,6 +1,8 @@
-package com.rapleaf.jack;
+package com.rapleaf.jack.queries;
 
 import java.util.Set;
+
+import com.rapleaf.jack.queries.Column;
 
 public interface Table {
