+++ src/java/com/rapleaf/jack/test_project/database_1/models/Image.java
@@ -13,10 +13,10 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-import com.rapleaf.jack.AbstractTable;
+import com.rapleaf.jack.queries.AbstractTable;
 import com.rapleaf.jack.AttributesWithId;
 import com.rapleaf.jack.BelongsToAssociation;
-import com.rapleaf.jack.Column;
+import com.rapleaf.jack.queries.Column;
 import com.rapleaf.jack.ModelIdWrapper;
 import com.rapleaf.jack.ModelWithId;
 import com.rapleaf.jack.test_project.IDatabases;
